Improvements to the ready 'rdy' commandline utility. 'rdy' now includes boost as a link dependency, and can do much more than it did before.
I will soon also be able to contribute several Houdini Digital Assets (HDA's) that will allow you (by pointing at this rdy binary) to import
OpenCLImageRD-based formulas directly into Houdini as Gas OpenCL Dops.

This is what the help (-h or --help argument) says in this version:

========================
Ready commandline utility that uses readybase as a processing (and data-retrieval) back-end.

It responds to various commandline arguments to load and print out parts of the system,
and can also still be used (like it used to work) to iterate the system to a specified number
of steps, and save it to a new vti file.

The various print options can facilitate the import of ready simulations into other applications, (such
as Houdini), without the need to actually link the ready libraries. This includes reagent initial-states,
(via -m), so be ready for some large lumps of text on stdout when using that argument.

Do let the Ready team know if there is something that you wish to print that currently isn't supported.

Options:
  -h [ --help ]                        Print the help message
  -n [ --num-iterations ] arg (=0)     Number of iterations to run before
                                       saving
  -k [ --print-kernel ]                Print (full) OpenCL kernel source (when
                                       possible)
  -f [ --print-formula ]               Print kernel formula (when possible)
  -r [ --print-reagent-info ]          Print reagent info
  -p [ --print-parameter-info ]        Print parameter info
  -s [ --print-render-settings ]       Print render Settings
  -d [ --print-formula-description ]   Print formula Description
  -m [ --print-initial-state-images ]  Print initial state images (Warning: May
                                       be large!)
  -i [ --vti-in ] arg                  VTI file to load (required)
  -o [ --vti-out ] arg                 VTI file to save (optional)
  -l [ --opencl-platform ] arg         OpenCL platform number (Currently will
                                       crash if incorrect!)
  -g [ --opencl-device ] arg           OpenCL device number (Currently will
                                       crash if incorrect!)
  -v [ --verbose ]                     Verbose output.
========================